V1.0.1

+++ app/src/main/java/in/vyomsoft/noti/task/TasksEntry/TasksEntryViewModel.kt
@@ -1,17 +1,19 @@
 package `in`.vyomsoft.noti.task.TasksEntry
 
 import TodoUiState
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
+import android.os.Bundle
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import `in`.vyomsoft.noti.GA4.AppAnalytics
 import `in`.vyomsoft.noti.apiUtils.Repository
 import `in`.vyomsoft.noti.requests.TodoRequest
-import `in`.vyomsoft.noti.responses.NotesRequest
-import `in`.vyomsoft.noti.responses.NotesResponse
 import `in`.vyomsoft.noti.responses.TodoResponse
+import `in`.vyomsoft.noti.utils.NetworkUtils
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.launch
 import okhttp3.ResponseBody
 import retrofit2.Call
 import retrofit2.Callback
@@ -24,8 +26,21 @@ class TasksEntryViewModel(private val repository: Repository) : ViewModel() {
     private val _updatedTodoResult = MutableStateFlow<TodoResponse?>(null)
     val updatedTodoResult: StateFlow<TodoResponse?> = _updatedTodoResult.asStateFlow()
 
-    private val _error = MutableLiveData<String?>()
-    val error: LiveData<String?> = _error
+    private val _isOnline = MutableStateFlow(true)
+    val isOnline: StateFlow<Boolean> = _isOnline.asStateFlow()
+
+    init {
+        monitorConnectivity()
+    }
+
+    private fun monitorConnectivity() {
+        viewModelScope.launch {
+            while (true) {
+                _isOnline.value = NetworkUtils.isInternetAvailable(repository.context)
+                delay(3000)
+            }
+        }
+    }
 
     fun saveTodo(request: TodoRequest, todoId: Long = -1L) {
         if (todoId == -1L || todoId == 0L) {
@@ -36,67 +51,73 @@ class TasksEntryViewModel(private val repository: Repository) : ViewModel() {
     }
 
     fun createTodo(todo: TodoRequest) {
+        if (!_isOnline.value) {
+            _uiState.value = TodoUiState.Error("No internet connection")
+            return
+        }
+
         _uiState.value = TodoUiState.Loading
         repository.createTodo(todo).enqueue(object : Callback<TodoResponse> {
-            override fun onResponse(
-                call: Call<TodoResponse>,
-                response: Response<TodoResponse>
-            ) {
+            override fun onResponse(call: Call<TodoResponse>, response: Response<TodoResponse>) {
                 if (response.isSuccessful) {
                     _uiState.value = TodoUiState.Success
-                    _updatedTodoResult.value = null
                     _updatedTodoResult.value = response.body()
+                    AppAnalytics.logEvent("Task_Created", Bundle().apply {
+                        putString("task_id", response.body()?.id.toString())
+                    })
                 } else {
-                    _error.postValue("Error Creating Todo")
+                    _uiState.value = TodoUiState.Error("Error Creating Todo")
                 }
             }
 
             override fun onFailure(call: Call<TodoResponse>, t: Throwable) {
-                _error.postValue("Error Creating Todo}")
+                _uiState.value = TodoUiState.Error("Check your Wi-Fi or cellular data and try again.")
             }
         })
     }
 
     fun getTask(taskId: Long) {
+        if (!_isOnline.value) {
+            _uiState.value = TodoUiState.Error("No internet connection")
+            return
+        }
+
         _uiState.value = TodoUiState.Loading
         repository.getTodo(taskId).enqueue(object : Callback<TodoResponse> {
-            override fun onResponse(
-                call: Call<TodoResponse>,
-                response: Response<TodoResponse>
-            ) {
+            override fun onResponse(call: Call<TodoResponse>, response: Response<TodoResponse>) {
                 if (response.isSuccessful) {
                     _updatedTodoResult.value = response.body()
                     _uiState.value = TodoUiState.Ready
                 } else {
-                    _error.postValue("Error Fetching Notes")
+                    _uiState.value = TodoUiState.Error("Error Fetching Task Group")
                 }
             }
 
             override fun onFailure(call: Call<TodoResponse>, t: Throwable) {
-                _error.postValue("Error Fetching Notes")
+                _uiState.value = TodoUiState.Error("Check your Wi-Fi or cellular data and try again.")
             }
         })
     }
 
     fun updateTodo(taskId: Long, todo: TodoRequest) {
+        if (!_isOnline.value) {
+            _uiState.value = TodoUiState.Error("No internet connection")
+            return
+        }
+
         _uiState.value = TodoUiState.Loading
         repository.updateTodo(todo, taskId).enqueue(object : Callback<TodoResponse> {
-            override fun onResponse(
-                call: Call<TodoResponse>,
-                response: Response<TodoResponse>
-            ) {
+            override fun onResponse(call: Call<TodoResponse>, response: Response<TodoResponse>) {
                 if (response.isSuccessful) {
                     _updatedTodoResult.value = response.body()
                     _uiState.value = TodoUiState.Success
                 } else {
-                    _error.postValue("Error Updating Task")
-                    _uiState.value = TodoUiState.Idle
+                    _uiState.value = TodoUiState.Error("Error Updating Task")
                 }
             }
 
             override fun onFailure(call: Call<TodoResponse>, t: Throwable) {
-                _error.postValue("Connection Failed: ${t.message}")
-                _uiState.value = TodoUiState.Idle
+                _uiState.value = TodoUiState.Error("Check your Wi-Fi or cellular data and try again.")
             }
         })
     }
@@ -110,30 +131,28 @@ class TasksEntryViewModel(private val repository: Repository) : ViewModel() {
     }
 
     fun deleteTodo(taskId: Int) {
-        _uiState.value = TodoUiState.Loading
+        if (!_isOnline.value) {
+            _uiState.value = TodoUiState.Error("No internet connection")
+            return
+        }
 
+        _uiState.value = TodoUiState.Loading
         repository.deleteTodo(taskId.toLong()).enqueue(object : Callback<ResponseBody> {
-            override fun onResponse(
-                call: Call<ResponseBody>,
-                response: Response<ResponseBody>
-            ) {
+            override fun onResponse(call: Call<ResponseBody>, response: Response<ResponseBody>) {
                 if (response.isSuccessful) {
                     _uiState.value = TodoUiState.Success
                 } else {
-                    _error.postValue("Error Deleting Task")
-                    _uiState.value = TodoUiState.Idle
+                    _uiState.value = TodoUiState.Error("Error Deleting Task")
                 }
             }
 
             override fun onFailure(call: Call<ResponseBody>, t: Throwable) {
-                _error.postValue("Connection Failed")
-                _uiState.value = TodoUiState.Idle
+                _uiState.value = TodoUiState.Error("Check your Wi-Fi or cellular data and try again.")
             }
         })
     }
 
     fun resetState() {
         _uiState.value = TodoUiState.Idle
-        _updatedTodoResult.value = null
     }
 }
